Extract Listing Card into Component
Extracted listing card into a component to be later re-used in other pages (for example home page).
