+++ frontend/vizzy/types/listing.ts
@@ -1,0 +1,8 @@
+export interface Listing {
+  id: string;
+  title: string;
+  type: 'sale' | 'rental' | 'giveaway' | 'swap';
+  price?: string;
+  pricePerDay?: string;
+  imageUrl: string;
+}
